Fix bug when a todo item is deleted both on the server and the client
If a todo item existed both on the server and the client, and then
deleted indepedently of each other on both the server and the client,
the old code would have a NULL local_vtodo but a non-NULL remote_vtodo.
Attempting to delete it from the database would result in an
AttributeError. Since the remote_vtodo is guaranteed to be non-NULL,
we will use that instead.