Only consider a merged todo updated if a field was modified
This fixes a bug where Nextcloud updates the SEQUENCE on save(),
which leads to Abeluna detecting it as a change on the server the next
sync.

However, no actual value has changed, leading to arbitrary saves
that don't change any value server-side.

Now, we only consider a merged todo updated only if a field has a
change.